Always return 64 bit integers (long) for numbers retrieved from the server. Convert old "(Integer)record.getValue()" to "record.getInt()" which performs the correct casting. Add utility methods to correctly cast numbers returned from user defined functions.

+++ examples/src/com/aerospike/examples/UserDefinedFunction.java
@@ -86,7 +86,7 @@ public class UserDefinedFunction extends Example {
 		client.put(params.writePolicy, key, bin);
 		
 		// Get record generation.
-		int gen = (Integer)client.execute(params.writePolicy, key, "record_example", "getGeneration");
+		long gen = (Long)client.execute(params.writePolicy, key, "record_example", "getGeneration");
 
 		// Write record if generation has not changed.
 		client.execute(params.writePolicy, key, "record_example", "writeIfGenerationNotChanged", Value.get(bin.name), bin.value, Value.get(gen));
